Update example for "Pitfalls in Race Condition Solutions" (#152)
The example code has a problem apart from the problem that is actually meant to be shown in this code. 
The withdraw function uses userBalance to get the amount to transfer, but then sets rewardsForA to 0, leaving userBalance untouched. Any user with a positive userBalance can this drain the contract simply by calling withdraw again and again - which is not the point this example was trying to make.
A withdraw (from userBalance) function should access userBalance only, while a withdrawReward function which is wanted here should only access rewardsForA. To clarify this, I renamed withdraw to withdrawReward.